% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fac.R
\name{create_fac}
\alias{create_fac}
\title{Create and upload a new fac}
\usage{
create_fac(facture_n = "DATENUM", client = "CLIENT",
  contact = "PERSONNE EN CHARGE", adresse = "LA BAS",
  objet = "Sponsoring", quantite = 1, prix = 1000,
  quoi = "SPONSORING DE", date = as.character(Sys.Date()),
  fait_a = "Paris", type = "sponsoring", dir = ".")
}
\arguments{
\item{facture_n}{generated by new_fac_num()}

\item{client}{client}

\item{contact}{contact chez le client}

\item{adresse}{adresse}

\item{objet}{objet}

\item{quantite}{quantite}

\item{prix}{prix}

\item{quoi}{quoi}

\item{date}{date}

\item{fait_a}{fait_a}

\item{type}{type}

\item{dir}{dir}
}
\description{
Create and upload a new fac
}
